Update patcher.rb
Added editBySub() data length check and user entry for windows mode instead of assuming the patcher is run on windows

+++ patcher.rb
@@ -1,41 +1,52 @@
+$windows_mode
+
 #modules
 module Patcher
 	#edits rom by substitution, maintaining exact file size
-	#takes three strings: filepath, and replaced_data, new_data which are converted to hex
+	#rom_filepath: string containing the path to the file which should already have been validated
+	#replaced_data: ascii string data to replace in the rom
+	#new_data: ascii string data to substitute into the rom
 	def self.editBySub(rom_filepath, replaced_data, new_data)
-		#add check to make sure replaced and new data are the same size?
-		data = nil
+		#checks to make sure replaced and new data are the same size (otherwise the rom becomes corrupted)
+		if replaced_data.length == new_data.length
+			data = nil
 
-		puts "reading..." #read data from file, stored in data
-		File.open(rom_filepath, "rb") do |file|
-		  data = file.read
+			puts "reading..." #read data from file, stored in data
+			File.open(rom_filepath, "rb") do |file|
+			  data = file.read
+			end
+
+			puts "unpacking..." #convert binary file data to hex, stored in hex_data
+			hex_data = data.unpack('H*')[0]
+
+			puts "patching..." #execute the replacement defined by function parameters, and store in patch_data
+			puts "(#{replaced_data.unpack('H*')[0]} --> #{new_data.unpack('H*')[0]})"
+			patch_data = hex_data.gsub(replaced_data.unpack('H*')[0], new_data.unpack('H*')[0])
+
+			puts "packing..." #converts modified hex file data back to binary, stored in bin_data
+			bin_data = [patch_data].pack('H*')
+
+			File.write('mid.bin', bin_data) #store bin_data [without this step ~1 KiB of data is lost?]
+			patchfile_name = "#{File.basename(rom_filepath, ".nds")}_patched.nds" #setup outfile name
+
+			if $windows_mode
+				puts "applying windows-specific fix..." #replace all 0d 0a pairs with 0a because of a windows newline quirk
+
+				File.open(patchfile_name, 'wb') do |converted|
+				  File.open('mid.bin', 'rb').each_line do |line|
+					converted << line.gsub("\r\n", "\n") # Replace CRLF with LF
+				  end
+				end
+			end
+
+			#rom_filepath_patched should now contain a file of equal size to rom_filepath
+			#with the changes instructed by replaced_data, new_data
+			patch_filechars = File.size(patchfile_name)
+			puts "fin: #{File.basename(patchfile_name)} [#{patch_filechars} bytes]"
+		else
+			puts "editBySub() received unusable data (different lengths error)"
+			puts "Process aborted"
 		end
-		
-		puts "unpacking..." #convert binary file data to hex, stored in hex_data
-		hex_data = data.unpack('H*')[0]
-		
-		puts "patching..." #execute the replacement defined by function parameters, and store in patch_data
-		puts "(#{replaced_data.unpack('H*')[0]} --> #{new_data.unpack('H*')[0]})"
-		patch_data = hex_data.gsub(replaced_data.unpack('H*')[0], new_data.unpack('H*')[0])
-		
-		puts "packing..." #converts modified hex file data back to binary, stored in bin_data
-		bin_data = [patch_data].pack('H*')
-		
-		File.write('mid.bin', bin_data) #store bin_data [without this step ~1 KiB of data is lost?]
-		patchfile_name = "#{File.basename(rom_filepath, ".nds")}_patched.nds" #setup outfile name
-		
-		puts "fixing..." #replace all 0d 0a pairs with 0a because of a windows newline quirk
-		
-		File.open(patchfile_name, 'wb') do |converted|
-		  File.open('mid.bin', 'rb').each_line do |line|
-			converted << line.gsub("\r\n", "\n") # Replace CRLF with LF
-		  end
-		end
-		
-		#rom_filepath_patched should now contain a file of equal size to rom_filepath
-		#with the changes instructed by replaced_data, new_data
-		patch_filechars = File.size(patchfile_name)
-		puts "fin: #{File.basename(patchfile_name)} [#{patch_filechars} bytes]"
 	end
 end
 
@@ -100,6 +111,20 @@ def help()
   menu()
 end
 
+def windows_check()
+	puts "Are you running this patcher on Windows? (Y)"
+	
+	entry = gets.chop
+	if entry == "y" || entry == "Y"
+		$windows_mode = true
+		puts "Running in windows mode..."
+	else
+		$windows_mode = false
+		puts "The patcher will not apply windows-specific fixes."
+	end
+end
+	
+
 #Main function
 def menu()
   #Display program options
@@ -130,4 +155,5 @@ end
 #program
 puts "Booted patcher."
 #add warnings?
+windows_check()
 menu()
